refactor(wrds_api): put wrds code into specific folder and put function to extract constituents in get_sp500_constituents.py
